fix PR comments refactor to have seperate modules instead of one big module add timing for fetch_url load OpenDTU test data from docs add constants
