+++ constants.py
@@ -1,0 +1,5 @@
+'''Global constants'''
+
+DTUVARIANT_AHOY = "ahoy"
+DTUVARIANT_OPENDTU = "opendtu"
+DTUVARIANT_TEMPLATE = "template"
